+++ packages/sdk/src/plt/index.ts
@@ -12,6 +12,7 @@ export * as TokenModuleReference from './TokenModuleReference.js';
 export * as TokenMetadataUrl from './TokenMetadataUrl.js';
 export * as TokenAmount from './TokenAmount.js';
 export * as Token from './Token.js';
+export * as Lock from './Lock.js';
 export * as Cbor from './Cbor.js';
 export * as CborMemo from './CborMemo.js';
 export * as CborAccountAddress from './CborAccountAddress.js';